Fixed errors in initial commit (Utils finished), removed some warnings

+++ Pokemon Multiplayer Red/TileMap.cpp
@@ -21,15 +21,15 @@ void TileMap::Draw(sf::RenderWindow* window, int dest_x, int dest_y, unsigned in
 	if (!tiles_tex)
 		return;
 	sf::IntRect src_rect = sf::IntRect(0, 0, 8, 8);
-	for (int y = 0; y < tile_size_y; y++)
+	for (unsigned int y = 0; y < tile_size_y; y++)
 	{
-		for (int x = 0; x < tile_size_x; x++)
+		for (unsigned int x = 0; x < tile_size_x; x++)
 		{
 			unsigned char t = (formation != 0 ? formation[tile * tile_size_x * tile_size_y + y * tile_size_x + x] : tile * tile_size_x * tile_size_y + y * tile_size_x + x);
 			src_rect.left = (t % tiles_x) * 8;
 			src_rect.top = (t / tiles_x) * 8;
 			sprite8x8.setTextureRect(src_rect);
-			sprite8x8.setPosition(dest_x * 8 * tile_size_x + x * 8, dest_y * 8 * tile_size_y + y * 8);
+			sprite8x8.setPosition((float)(dest_x * 8 * tile_size_x + x * 8), (float)(dest_y * 8 * tile_size_y + y * 8));
 			window->draw(sprite8x8);
 		}
 	}
